+++ apps/io-services-cms-webapp/src/lib/clients/jira-legacy-client.ts
@@ -5,14 +5,12 @@ import * as E from "fp-ts/lib/Either";
 import * as O from "fp-ts/lib/Option";
 import * as TE from "fp-ts/lib/TaskEither";
 import { pipe } from "fp-ts/lib/function";
+import { isNumber } from "fp-ts/lib/number";
+import * as T from "fp-ts/Task";
 import * as t from "io-ts";
 import nodeFetch from "node-fetch-commonjs";
 import { IConfig } from "../../config";
-import {
-  JIRA_REST_API_PATH,
-  SearchJiraIssuesPayload,
-  checkJiraResponse,
-} from "./jira-client";
+import { JIRA_REST_API_PATH, SearchJiraIssuesPayload } from "./jira-client";
 
 export const JIRA_SERVICE_TAG_PREFIX = "devportal-service-";
 
@@ -28,15 +26,19 @@ export type JiraLegacyIssueStatus = t.TypeOf<typeof JiraLegacyIssueStatus>;
 export const JiraLegacyIssue = t.type({
   id: NonEmptyString,
   key: NonEmptyString,
-  fields: t.type({
-    comment: t.type({
-      comments: t.readonlyArray(t.type({ body: t.string })),
+  fields: t.intersection([
+    t.type({
+      comment: t.type({
+        comments: t.readonlyArray(t.type({ body: t.string })),
+      }),
+      status: t.type({
+        name: JiraLegacyIssueStatus,
+      }),
     }),
-    status: t.type({
-      name: JiraLegacyIssueStatus,
+    t.partial({
+      statuscategorychangedate: t.string,
     }),
-    statuscategorychangedate: NonEmptyString,
-  }),
+  ]),
 });
 export type JiraLegacyIssue = t.TypeOf<typeof JiraLegacyIssue>;
 
@@ -54,9 +56,15 @@ export type SearchJiraLegacyIssuesResponse = t.TypeOf<
   typeof SearchJiraLegacyIssuesResponse
 >;
 
+export type JiraLegacyOptions = {
+  defaultRetryAfter?: number;
+  maxRetry?: number;
+};
+
 export type JiraLegacyAPIClient = {
   readonly searchJiraIssueByServiceId: (
-    serviceId: ServiceId
+    serviceId: ServiceId,
+    options?: JiraLegacyOptions
   ) => TE.TaskEither<Error, O.Option<JiraLegacyIssue>>;
 };
 
@@ -72,7 +80,16 @@ export const jiraLegacyClient = (
     "Content-Type": "application/json",
   };
 
-  const searchJiraIssueByServiceId = (serviceId: ServiceId) => {
+  const searchJiraIssueByServiceId = (
+    serviceId: ServiceId,
+    {
+      defaultRetryAfter = 5000,
+      maxRetry = 5,
+    }: {
+      defaultRetryAfter?: number;
+      maxRetry?: number;
+    } = {}
+  ) => {
     const bodyData: SearchJiraIssuesPayload = {
       fields: [
         "summary",
@@ -89,17 +106,38 @@ export const jiraLegacyClient = (
     };
 
     return pipe(
-      TE.tryCatch(
-        () =>
-          fetchApi(`${config.JIRA_NAMESPACE_URL}${JIRA_REST_API_PATH}search`, {
-            body: JSON.stringify(bodyData),
-            headers: jiraHeaders,
-            method: "POST",
-          }),
-        E.toError
+      fetchApiWithRetry(
+        `${config.JIRA_NAMESPACE_URL}${JIRA_REST_API_PATH}search`,
+        fetchApi,
+        {
+          body: JSON.stringify(bodyData),
+          headers: jiraHeaders,
+          method: "POST",
+        },
+        defaultRetryAfter,
+        maxRetry
       ),
-      TE.chainEitherK(checkJiraResponse),
-      TE.chain((response) => TE.tryCatch(() => response.json(), E.toError)),
+      TE.chain((response) =>
+        pipe(
+          TE.tryCatch(
+            () => response.json(),
+            (err) =>
+              new Error(
+                `Error parsing Jira response, statusCode: ${
+                  response.status
+                }, headers: ${extractHeaders(response)}, error: ${
+                  E.toError(err).message
+                }`
+              )
+          ),
+          TE.chain((responseBody) =>
+            pipe(
+              checkJiraResponse(response.status, responseBody),
+              TE.fromEither
+            )
+          )
+        )
+      ),
       TE.chain((responseBody) =>
         pipe(
           responseBody,
@@ -110,6 +148,104 @@ export const jiraLegacyClient = (
         )
       )
     );
+  };
+
+  const fetchApiWithRetry = (
+    url: string,
+    fetchApi: typeof fetch,
+    options: RequestInit,
+    defaultRetryAfter: number,
+    maxRetryCount: number
+  ): TE.TaskEither<Error, Response> => {
+    const attempt = (retryCount: number): TE.TaskEither<Error, Response> =>
+      pipe(
+        TE.tryCatch(() => fetchApi(url, options), E.toError),
+        TE.chain((response) =>
+          response.status === 429
+            ? pipe(
+                response.headers.get("Retry-After"),
+                O.fromNullable,
+                O.chain((delay) =>
+                  isNumber(Number(delay)) ? O.some(delay) : O.none
+                ),
+                O.fold(
+                  () => defaultRetryAfter,
+                  (retryAfter) => Number(retryAfter) * 1000
+                ),
+                (retryAfter) =>
+                  retryCount < maxRetryCount
+                    ? pipe(
+                        T.delay(retryAfter)(attempt(retryCount + 1)),
+                        TE.map(TE.right),
+                        TE.flatten
+                      )
+                    : TE.left(
+                        new Error(
+                          `Cannot Contact jira after retries, statusCode: ${
+                            response.status
+                          }, headers: ${extractHeaders(response)}`
+                        )
+                      )
+              )
+            : TE.right(response)
+        )
+      );
+
+    return attempt(0);
+  };
+
+  const extractHeaders = (response: Response) => {
+    try {
+      const headers: Record<string, string> = {};
+      response.headers.forEach((value, key) => {
+        // eslint-disable-next-line functional/immutable-data
+        headers[key] = value;
+      });
+      return JSON.stringify(headers);
+    } catch (e) {
+      return "CANNOT EXTRACT HEADERS";
+    }
+  };
+
+  const checkJiraResponse = (
+    status: number,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    responseBody: any
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  ): E.Either<Error, any> => {
+    if (status === 200 || status === 201) {
+      return E.right(responseBody);
+    } else if (status === 400) {
+      return E.left(
+        new Error(
+          `Invalid request, responseBody: ${JSON.stringify(responseBody)}`
+        )
+      );
+    } else if (status === 401) {
+      return E.left(
+        new Error(
+          `Jira secrets misconfiguration, responseBody: ${JSON.stringify(
+            responseBody
+          )}`
+        )
+      );
+    } else if (status >= 500) {
+      return E.left(
+        new Error(
+          `Jira API returns an error, responseBody: ${JSON.stringify(
+            responseBody
+          )}`
+        )
+      );
+    } else {
+      return E.left(
+        new Error(
+          `Unknown status code ${status} received, responseBody: ${JSON.stringify(
+            responseBody
+          )}`
+        )
+      );
+    }
   };
 
   return {