+++ apps/io-services-cms-webapp/src/watchers/on-legacy-service-change.ts
@@ -138,7 +138,17 @@ const legacyToCms = (
   item: Service
 ) =>
   pipe(
-    jiraLegacyClient.searchJiraIssueByServiceId(item.serviceId),
+    item,
+    O.fromPredicate(
+      (item) =>
+        !isDeletedService(item) &&
+        isValidService(qualityCheckExclusionList)(item) &&
+        !item.isVisible
+    ),
+    O.fold(
+      () => TE.right(O.none), // evitiamo la chiamata a Jira quando non serve, simulando una sua risposta vuota
+      (_) => jiraLegacyClient.searchJiraIssueByServiceId(item.serviceId)
+    ),
     TE.map(
       flow(
         O.fold(
@@ -187,7 +197,7 @@ export const handler =
                   TE.mapLeft(
                     (e) =>
                       new Error(
-                        `Error while processing serviceId ${item.serviceId}, the reason was => ${e.message}`
+                        `Error while processing serviceId ${item.serviceId}, the reason was => ${e.message}, the stack was => ${e.stack}`
                       )
                   )
                 )